fix(db): align run lease fence schema
Migration 0002 declared agent_run_leases.fence as sa.BigInteger() while the
ORM model (AgentRunLeaseModel.fence) declares mapped_column(Integer). The
mismatch broke 'alembic check' (BIGINT <-> Integer drift), which blocked
'alembic upgrade head'/'alembic check' and therefore the entire postgres job
(PG concurrency suite never ran).

0002 is not yet in main, so the cleanest fix is to align the unmerged
migration to the ORM (the authoritative source of column types): fence is a
monotonic fencing token (1,2,3...); Integer (INT4) is more than sufficient and
matches the model. No new fix-migration added.

+++ alembic/versions/20260820_0002_add_concurrency_guards.py
@@ -49,7 +49,7 @@ def upgrade() -> None:
         ),
         sa.Column(
             "fence",
-            sa.BigInteger(),
+            sa.Integer(),
             nullable=False,
             server_default="1",
             comment="fencing token：每次抢占 +1，旧 worker 凭旧 fence 被拒绝",
